rename enum to extended_enum to avoid collisions with python builtin module
