+++ src/scooze/catalogs.py
@@ -2,7 +2,7 @@ from enum import StrEnum, auto
 from functools import cache
 from typing import FrozenSet
 
-from scooze.enum import ExtendedEnum
+from scooze.extended_enum import ExtendedEnum
 
 # region Card Enums
