+++ src/scooze/extended_enum.py
