+++ src/scooze/models/card.py
@@ -36,7 +36,7 @@ from scooze.catalogs import (
     SecurityStamp,
     SetType,
 )
-from scooze.enum import DbCollection
+from scooze.extended_enum import DbCollection
 from scooze.models.cardparts import (
     CardFaceModel,
     ImageUrisModel,
