+++ src/scooze/models/deck.py
@@ -3,7 +3,7 @@ from datetime import date
 
 from pydantic import ConfigDict, Field, field_serializer, model_validator
 from scooze.catalogs import Format
-from scooze.enum import DbCollection
+from scooze.extended_enum import DbCollection
 from scooze.models.utils import ObjectIdT, ScoozeBaseModel, ScoozeDocument
 from scooze.utils import cmdr_size, encode_date, main_size, side_size
