+++ src/scooze/utils.py
@@ -12,7 +12,7 @@ from frozendict import frozendict
 from pydantic.alias_generators import to_camel
 from scooze.catalogs import CostSymbol, Format
 from scooze.config import CONFIG
-from scooze.enum import ExtendedEnum
+from scooze.extended_enum import ExtendedEnum
 
 ## Generic Types
 T = TypeVar("T")  # generic type
